+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 * Added "email profile" to scope in `Assent.Strategy.AzureAD`
 * Use `response_mode=form_post` for `Assent.Strategy.AzureAD`
 * Updated `Assent.Strategy.OAuth2` to handle access token request correctly when `:auth_method` is `nil` per RFC specs
+* Remove overriden
+* Changed `Assent.Strategy.Apple` to use OIDC strategy and verify the JWT
 
 ## v0.1.4 (2019-11-09)